Refactor movie entry point
