+++ main.rb
@@ -1,60 +1,10 @@
 require_relative 'class/item'
-require_relative 'class/movie'
-require_relative 'class/source'
 require_relative 'class/music_album'
 require_relative 'class/genre'
-require_relative 'app'
+require_relative 'apps/game_app'
+require_relative 'apps/movie_app'
 require 'json'
 
-# Load existing data from MOVIES JSON files (if any)
-def movies_load_data
-  movies = []
-  if File.exist?('data/movies.json')
-    File.open('data/movies.json', 'r') do |file|
-      file.each_line do |line|
-        movies << Movie.from_json(line)
-      end
-    end
-  end
-  movies
-end
-
-# Save data to MOVIES JSON files
-def movies_save_data(movies)
-  File.open('data/movies.json', 'w') do |file|
-    movies.each do |item|
-      file.puts JSON.dump(item.to_json)
-    end
-  end
-end
-
-# Initialize movies
-movies = movies_load_data
-
-# Load existing data from MUSIC_ALBUMS JSON files (if any)
-def music_load_data
-  music_albums = []
-  if File.exist?('data/music_album.json')
-    File.open('data/music_album.json', 'r') do |file|
-      file.each_line do |line|
-        music_albums << MusicAlbum.from_json(line)
-      end
-    end
-  end
-  music_albums
-end
-
-# Save data to MUSIC_ALBUMS JSON files
-def music_save_data(music_albums)
-  File.open('data/music_album.json', 'w') do |file|
-    music_albums.each do |item|
-      file.puts JSON.dump(item.to_json)
-    end
-  end
-end
-
-# Initialize music_albums
-music_albums = music_load_data
 
 def display_menu
   puts '------------------------------------------------------'
@@ -75,11 +25,12 @@ def display_menu
   puts '13 - Add a genre'
   puts '14 - Add a label'
   puts '15 - Add a author'
-  puts '16 - Add a source'
-  puts '17 - Exit'
+  puts '16 - Exit'
 end
 
-app = App.new
+
+movie_app = MovieApp.new
+app = GameApp.new
 loop do
   display_menu
   option = gets.chomp.to_i
@@ -94,10 +45,8 @@ loop do
     end
   when 3
     # List All Movies
-    puts 'List of Movies:'
-    movies.each do |item|
-      puts "- Title: #{item.title}, Source: #{item.source}, Published: #{item.publish_date}, Archived: #{item.archived}"
-    end
+    #movie_app.movies_load_data
+    movie_app.list_movies
   when 4
     app.list_games
   when 5
@@ -113,11 +62,7 @@ loop do
     app.list_authors
   when 8
     # List All Sources
-    puts 'List of Sources:'
-    sources = movies.map(&:source).uniq.compact
-    sources.each do |source|
-      puts "- Name: #{source}"
-    end
+    movie_app.list_movie_source
   when 9
     # your code here
   when 10
@@ -136,18 +81,7 @@ loop do
     puts 'Music_album added!'
   when 11
     # Add A Movie
-    puts 'Enter movie title:'
-    title = gets.chomp
-    puts 'Enter publish date (YYYY-MM-DD):'
-    publish_date = gets.chomp
-    puts 'Is it silent? (y/n)'
-    silent = true if gets.chomp.downcase == 'y'
-
-    movie = Movie.new(title, publish_date, silent: silent)
-    movie.move_to_archive
-    movies << movie
-
-    puts 'Movie added!'
+   movie_app.add_movie
   when 12
     app.add_game
   when 13
@@ -166,21 +100,10 @@ loop do
   when 15
     app.add_author
   when 16
-    # Add Movie Source
-    puts 'Enter source name:'
-    source_name = gets.chomp
-
-    source = Source.new(source_name)
-    movies.each do |item|
-      item.source = source.name if item.source.nil?
-    end
-
-    puts 'Source added!'
-  when 17
     # Save Movies data to JSON files before exiting
-    movies_save_data(movies)
+    # movie_app.movies_save_data(movies)
     # Save Music_album data to JSON files before exiting
-    music_save_data(music_albums)
+    # music_save_data(music_albums)
     puts 'Thanks for using Catalog of My Things app!'
     break
   else
